+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-N/A
+### Fixed
+
+- `QueryManager` accepts `translations`, enabling it to produce correct placeholder values for `fields`, `operators`, and `values` lists.
 
 ## [v8.22.1] - 2026-08-04